Implemented a four vector
Despite the fact that I'm doing 3D graphics, I'm going to work in 4D.
The reason for it is that then I can represent translations in 3D with
the same matrix machinery that I would use for scaling and rotations.

The four_vector is implemented using avx2 vectors. This means that
additions, subtractions, elementvise mutliplication and division are all
single instructions.