Interface for the four_matrix
I'm going to use four matrices to represent all geometric
transformations. So far, rotations in the xy-, xz- and yz-planes, are
implemented, and the identity matrix. Furthermore a start to do a
general rotation a round an arbitrary axis has been created.

Furthermore, interfaces for applying the transformations on to vectors
has been written and a matrix matrix multiplication, to combine
transformations.